refactor(test): migrate compose helpers to three-tier structure
Split api-test-helpers/agents.ts into three tiers:
- db-test-seeders/agents.ts: impossible-state seeders with @why-db-direct docs
- db-test-assertions/agents.ts: read-only query helpers
- api-test-helpers/agents.ts: API-based helpers (retained)

Updated imports across ~30 test files. Migrated skills, slack/events,
export, and backfill tests to use appropriate tier. Removed 15 functions
from api-test-helpers that were moved to seeders/assertions.

Closes #9340
