+++ turbo/apps/web/app/api/internal/callbacks/voice-chat-prepare/__tests__/route.test.ts
@@ -9,10 +9,10 @@ import {
   createTestCallback,
   createTestRunInDb,
   createSignedCallbackRequest,
-  getTestZeroAgentId,
   insertTestVoiceChatPreparation,
   getTestVoiceChatPreparation,
 } from "../../../../../../src/__tests__/api-test-helpers";
+import { getTestZeroAgentId } from "../../../../../../src/__tests__/db-test-assertions/agents";
 import { POST } from "../route";
 
 const context = testContext();
