+++ turbo/apps/web/app/api/zero/agents/[id]/user-connectors/__tests__/route.test.ts
@@ -4,9 +4,9 @@ import { POST as postAgentRoute } from "../../../route";
 import {
   createTestRequest,
   createTestCliToken,
-  setComposeHeadVersion,
-  getComposeHeadVersion,
 } from "../../../../../../../src/__tests__/api-test-helpers";
+import { setComposeHeadVersion } from "../../../../../../../src/__tests__/db-test-seeders/agents";
+import { getComposeHeadVersion } from "../../../../../../../src/__tests__/db-test-assertions/agents";
 import {
   testContext,
   type UserContext,
