+++ turbo/apps/web/app/api/zero/chat/messages/__tests__/route.test.ts
@@ -4,12 +4,12 @@ import { POST } from "../route";
 import {
   createTestRequest,
   createTestCompose,
-  getTestZeroAgentId,
   insertOrgDefaultModelProvider,
   findTestCallbacksByRunId,
   findTestRunRecord,
   getTestRun,
 } from "../../../../../../src/__tests__/api-test-helpers";
+import { getTestZeroAgentId } from "../../../../../../src/__tests__/db-test-assertions/agents";
 import { POST as createChatThreadPOST } from "../../../chat-threads/route";
 import {
   testContext,
