+++ turbo/apps/web/app/api/zero/integrations/slack/message/__tests__/route.test.ts
@@ -5,12 +5,12 @@ import {
   createTestRequest,
   createTestCompose,
   createTestRunInDb,
-  createTestZeroAgent,
   createTestSlackOrgInstallation,
   insertOrgMembersCacheEntry,
   createTestSchedule,
   seedTestSlackOrgConnection,
 } from "../../../../../../../src/__tests__/api-test-helpers";
+import { createTestZeroAgent } from "../../../../../../../src/__tests__/db-test-seeders/agents";
 import {
   testContext,
   uniqueId,
