+++ turbo/apps/web/app/api/zero/logs/__tests__/route.test.ts
@@ -6,11 +6,11 @@ import {
   createTestCompose,
   createTestRunInDb,
   createTestSchedule,
-  createTestZeroAgent,
   createTestRun,
   completeTestRun,
   insertOrgMembersCacheEntry,
 } from "../../../../../src/__tests__/api-test-helpers";
+import { createTestZeroAgent } from "../../../../../src/__tests__/db-test-seeders/agents";
 import {
   testContext,
   type UserContext,
