+++ turbo/apps/web/app/api/zero/schedules/[name]/disable/__tests__/route.test.ts
@@ -3,13 +3,13 @@ import { POST } from "../route";
 import {
   createTestRequest,
   createTestCompose,
-  createTestZeroAgent,
-  getTestZeroAgentId,
   createTestOrg,
   createTestSchedule,
   enableTestSchedule,
   getTestSchedule,
 } from "../../../../../../../src/__tests__/api-test-helpers";
+import { createTestZeroAgent } from "../../../../../../../src/__tests__/db-test-seeders/agents";
+import { getTestZeroAgentId } from "../../../../../../../src/__tests__/db-test-assertions/agents";
 import {
   testContext,
   uniqueId,
